requirements.txt
adal==1.2.7
aiohttp==3.9.5
aiosignal==1.3.1
amqp==5.1.1
anyio==3.6.2
appnope==0.1.3
arabic-reshaper==3.0.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.6.0
asn1crypto==1.5.1
astroid==2.11.7
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
Authlib==1.2.1
azure-common==1.1.0
azure-core==1.26.4
azure-identity==1.7.1
azure-mgmt-core==1.4.0
azure-mgmt-media==8.0.0
azure-nspkg==3.0.2
azure-storage-blob==12.9.0
backcall==0.2.0
billiard==3.6.4.0
black==23.3.0
bleach==6.0.0
Brotli==1.0.9
celery==5.2.7
certifi==2023.5.7
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coverage==7.2.2
cron-descriptor==1.3.0
cryptography==40.0.2
cssselect2==0.7.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.6
distlib==0.3.6
Django==4.0.10
django-anymail==9.1
django-braces==1.15.0
django-celery-beat==2.5.0
django-ckeditor==6.7.0
django-cors-headers==3.14.0
django-coverage-plugin==3.0.0
django-crontab==0.7.1
django-environ==0.10.0
django-extensions==3.2.1
django-filter==23.2
django-js-asset==2.1.0
django-oauth-toolkit==2.3.0
django-phonenumber-field==6.1.0
django-redis==5.2.0
django-rest-framework-social-oauth2==1.1.0
django-storages==1.13.2
django-timezone-field==5.0
django-user-agents==0.4.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
executing==1.2.0
Faker==18.7.0
filelock==3.12.0
flake8==6.0.0
flake8-isort==6.0.0
flower==1.2.0
fonttools==4.39.4
frozenlist==1.4.1
geoip2==4.8.0
gunicorn==20.1.0
h11==0.14.0
hiredis==2.2.2
html5lib==1.1
humanize==4.6.0
icalendar==5.0.12
identify==2.5.24
idna==3.4
inflect==7.2.0
iniconfig==2.0.0
ipdb==0.13.13
ipython==8.13.2
isodate==0.6.1
isort==5.10.1
jedi==0.18.2
jwcrypto==1.5.0
kombu==5.2.4
lazy-object-proxy==1.9.0
# lxml==4.9.2
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
maxminddb==2.6.1
mccabe==0.7.0
more-itertools==10.2.0
msal==1.22.0
msal-extensions==0.3.1
msrest==0.7.1
msrestazure==0.6.4
mssql-django==1.2
multidict==6.0.5
mypy-extensions==1.0.0
nodeenv==1.8.0
# numpy==1.25.0
oauthlib==3.2.2
openpyxl==3.1.2
oscrypto==1.3.0
packaging==23.1
pandas==2.0.2
parso==0.8.3
pathspec==0.11.1
pexpect==4.8.0
phonenumbers==8.13.11
pickleshare==0.7.5
Pillow==9.5.0
pipenv==2023.4.20
platformdirs==3.5.1
pluggy==1.0.0
portalocker==2.7.0
pre-commit==3.2.1
prometheus-client==0.16.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
pycparser==2.21
pydyf==0.6.0
pyflakes==3.0.1
Pygments==2.15.1
pyHanko==0.18.1
pyhanko-certvalidator==0.22.0
PyJWT==2.7.0
pylint==2.13.5
pylint-celery==0.3
pylint-django==2.5.3
pylint-plugin-utils==0.8.1
pyodbc==5.1.0
pypdf==3.9.0
pyphen==0.14.0
pypng==0.20220715.0
pytest==7.2.2
pytest-django==4.5.2
python-bidi==0.4.2
python-crontab==2.7.1
python-dateutil==2.8.2
python-http-client==3.3.7
python-slugify==8.0.1
python3-openid==3.2.0
pytz==2023.3
# PyYAML==6.0
qrcode==7.4.2
razorpay==1.3.0
redis==4.5.4
reportlab==3.6.13
requests==2.30.0
requests-oauthlib==1.3.1
sendgrid==6.10.0
sentry-sdk==1.17.0
six==1.16.0
sniffio==1.3.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
stack-data==0.6.2
starkbank-ecdsa==2.2.0
svglib==1.5.1
text-unidecode==1.3
tinycss2==1.2.1
tornado==6.3.2
traitlets==5.9.0
typeguard==4.2.1
typing_extensions==4.11.0
tzdata==2023.3
tzlocal==5.0.1
ua-parser==0.16.1
uritools==4.0.1
urllib3==2.0.2
user-agents==2.2.0
uvicorn==0.21.1
vine==5.0.0
virtualenv==20.23.0
virtualenv-clone==0.5.7
watchfiles==0.19.0
watchgod==0.7
wcwidth==0.2.6
weasyprint==59.0
webencodings==0.5.1
Werkzeug==2.2.3
wrapt==1.15.0
xhtml2pdf==0.2.11
# yarl==1.9.4
# zopfli==0.2.2
